+++ src/Components/Awards.js
@@ -1,0 +1,16 @@
+import React, { Component } from 'react';
+
+class Awards extends Component {
+    render() {
+        return (
+            <section id="awards">
+                <div>
+                    <h1>Awards</h1>
+                    <p>samplkedpemdfoasnfoasnfansi</p>
+                </div>
+            </section>
+        );
+    }
+}
+
+export default Awards;